* gcutter/gcut-string-io-channel.c: use gint64 for offset.
git-svn-id: https://cutter.svn.sourceforge.net/svnroot/cutter/cutter/trunk@3191 10a583ed-1e3d-0410-b285-c80daaaae891

+++ gcutter/gcut-string-io-channel.c
@@ -33,8 +33,8 @@ struct _GCutStringIOChannel
 {
     GIOChannel channel;
     GString *string;
-    gsize write_offset;
-    gsize read_offset;
+    gint64 write_offset;
+    gint64 read_offset;
     gsize buffer_limit;
     gsize limit;
     gboolean read_fail;
